TLRC-284 Create 'Your T Levels' page   Api  and Web - Navigation Objects are mapped and Id param replaced with UkPrn

+++ src/Sfa.Tl.ResultsAndCertification.Application/Interfaces/IAwardingOrganisationService.cs
@@ -6,6 +6,6 @@ namespace Sfa.Tl.ResultsAndCertification.Application.Services.Interfaces
 {
     public interface IAwardingOrganisationService
     {
-        Task<IEnumerable<AwardingOrganisationPathwayStatus>> GetAllTlevelsByAwardingOrganisationIdAsync(int id);
+        Task<IEnumerable<AwardingOrganisationPathwayStatus>> GetAllTlevelsByAwardingOrganisationIdAsync(long id);
     }
 }
